ConnectionServiceErrorCodes moved into right module
